// frontend/src/app/api/model/learningCardRequest.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LearningStatus } from './learningStatus';


export interface LearningCardRequest { 
    title: string;
    description?: string | null;
    status?: LearningStatus;
}
export namespace LearningCardRequest {
}